+++ app/src/main/java/com/example/nytimesapp/domain/critics/CriticItemImg.kt
@@ -1,7 +1,7 @@
 package com.example.nytimesapp.domain.critics
 
 data class CriticItemImg(
-    val id:Float,
+    val id:Long,
     val srcRemote:String,
     val srcLocal:String,
     val imgHeight:Int,
