+++ app/src/main/java/com/example/nytimesapp/domain/reviews/ReviewItemLink.kt
@@ -1,7 +1,7 @@
 package com.example.nytimesapp.domain.reviews
 
 data class ReviewItemLink(
-    val id:Float,
+    val id:Long,
     val url:String,
     val suggestedLinkText:String?,
 )
